moving the logic of the delete() method to the service

+++ src/main/java/com/dream/team/library/service/impl/AuthorServiceImpl.java
@@ -83,7 +83,9 @@ public class AuthorServiceImpl implements AuthorService {
         log.info("Deleted the author");
         log.debug("Author: " + authorDto);
 
-        authorRepository.delete(AuthorMapper.INSTANCE.toAuthor(authorDto));
+        if (authorDto != null && authorDto.getId() != null) {
+            authorRepository.delete(AuthorMapper.INSTANCE.toAuthor(authorDto));
+        }
     }
 
     @Override
